tech item - code clean
