+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.1.0] - 26.11.2023
+- Websocket reconnection
+- Updated websocket-client version to 1.6.4
+- Analog sensor integration
+- Switch integration
+- Fix for entities received after unsupported partition resource
+- typos code quality ...
+
 ## [1.0.1] - 07.09.2023
 
 ### Added
