+++ custom_components/nexo/nexoEntities/nexoGate.py
@@ -1,0 +1,7 @@
+from nexoResource import NexoResource
+
+
+class NexoGate(NexoResource):
+    def __init__(self, web_socket, *args, **kwargs) -> None:
+        super().__init__(web_socket, *args, **kwargs)
+
